Created HTML Boilerplate for Recipe Page
Created Boilerplate to lay the foundation for the rest of the HTML
website that we are creating for the ODIN recipes project. Trying to
create more commits than usual, but I figured that the boilerplate is so
normal that it is probably not necessary to create a commit for each
part of it.